--- src/components/App.jsx
@@ -1,15 +1,0 @@
-import React from "react";
-import Main from "./main.jsx";
-
-const App = (props) => {
-  // eslint-disable-next-line react/prop-types
-  const {offersCount} = props;
-
-  return (
-    <Main
-      offersCount = {offersCount}
-    />
-  );
-};
-
-export default App;
